+++ wigglystuff/__init__.py
@@ -5,6 +5,7 @@ import importlib.metadata
 from .annotation import AnnotationWidget
 from .altair_widget import AltairWidget
 from .api_doc import ApiDoc
+from .bezier_curve import BezierCurve
 from .cell_tour import CellTour
 from .chart_puck import ChartPuck
 from .chart_multi_select import ChartMultiSelect
@@ -46,6 +47,7 @@ __all__ = [
     "AnnotationWidget",
     "AltairWidget",
     "ApiDoc",
+    "BezierCurve",
     "CellTour",
     "ChartPuck",
     "ChartMultiSelect",